WIP renaming limiters to setpoints
