+++ config/config.example.json
@@ -1,5 +1,5 @@
 {
-    "limiters": {
+    "setpoints": {
     },
     "inverters": [
         {
